requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-core
langchain-community